+++ src/lib/services/wasm/code/index.ts
@@ -95,3 +95,4 @@ export const useCodesByAddress = (
 };
 
 export * from "./gql";
+export * from "./lcd";